edge-ai copy/requirements.txt
fastapi
pillow
pytesseract
uvicorn
